lib: export access-lists shared functions
These functions are going to be used by the new filter northbound
implementation.
